+++ .gitignore
@@ -4,3 +4,4 @@
 /test
 blend_script.py
 run.bat
+notes.txt
